fix: reminderTimeFormatter 접근 범위를 public으로 수정
